Corrected forgotten double modification in GappDebugger
git-svn-id: svn+ssh://erebor.esa.informatik.tu-darmstadt.de/home/wimi/svn/GaalopCompiler/gaalop-2.0/branches/tbaAndGapp@1072 4235ee25-d3ef-403b-b6ed-692c26853d4b

+++ gappDebugger/src/main/java/de/gaalop/gappDebugger/Controller.java
@@ -41,7 +41,7 @@ import org.antlr.runtime.tree.CommonTreeNodeStream;
 public class Controller {
 
     private UI ui;
-    private HashMap<String, Float> inputValues = new HashMap<String, Float>();
+    private HashMap<String, Double> inputValues = new HashMap<String, Double>();
 
     private DefaultListModel modelVars = new DefaultListModel();
     private DefaultListModel modelSrc = new DefaultListModel();
@@ -154,7 +154,7 @@ public class Controller {
                 if (part.contains("=")) {
                     String[] parts = part.split("=");
                     String varname = parts[0].trim();
-                    Float value = Float.parseFloat(parts[1].trim());
+                    Double value = Double.parseDouble(parts[1].trim());
                     inputValues.put(varname, value);
                     changed = true;
                 }
